requirements.txt
requests~=2.25.0
Flask~=1.1.2
Pillow==8.2.0